requirements.txt
Pillow==9.2.0
pynput==1.7.6
pystray==0.19.4